refactor(frontend): reorganize frontend structure
Author: Rudy Savini - @SavGRY
