+++ frontend/incassi/src/app/app.config.ts
@@ -4,8 +4,8 @@ import {provideRouter} from '@angular/router';
 import {provideHttpClient, withInterceptors} from '@angular/common/http';
 import {provideAnimations} from '@angular/platform-browser/animations';
 import {provideAnimationsAsync} from '@angular/platform-browser/animations/async';
+import {authInterceptor} from '../shared/interceptor/auth.interceptor';
 import {routes} from './app.routes';
-import {authInterceptor} from './interceptor/auth.interceptor';
 
 export const appConfig: ApplicationConfig = {
   providers: [
