+++ frontend/incassi/src/models/Client.ts
@@ -1,0 +1,7 @@
+export interface Client {
+  code: number;
+  name: string;
+  address: string;
+  city: string;
+  province: string;
+}
